+++ bank_resource_server/src/main/java/ru/gb/group5984/model/transactions/Transaction.java
@@ -2,6 +2,8 @@ package ru.gb.group5984.model.transactions;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 
 /**
  * Данные о транзакции оплаты товаров:
@@ -13,5 +15,5 @@ import lombok.Data;
 public class Transaction {
     private Long creditAccount;
     private Long debitAccount;
-    private Double transferAmount;
+    private BigDecimal transferAmount;
 }
