+++ store_resource_server/src/main/java/ru/gb/group5984/model/basket/BasketInfo.java
@@ -2,6 +2,8 @@ package ru.gb.group5984.model.basket;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 /**
  * Информационная часть страницы о списке товаров, находящихся
  * в корзине покупателей.
@@ -13,5 +15,5 @@ public class BasketInfo {
     private Integer next;
     private Integer current;
     private Integer prev;
-    private Double totalPrice;
+    private BigDecimal totalPrice;
 }
